feat(vs2022_third): C4018 fix signed/unsigned mismatch — 11 files (batch 1)
Cast int loop variables to size_t/DWORD in comparisons:
- ConfigFile/IniFile.cpp: 3 instances (int→DWORD)
- NEXIO/NEXIO.CPP: 1 instance (int→size_t)
- Stage/STAGE.CPP: 1 instance (int→size_t)
- MeaSys/Xmp.cpp: 2 instances (int→size_t)
- NanoSpec/BCMenu.cpp: 1 instance (int→size_t)
- NanoSpec/ChifTransiStress.cpp: 6 instances (int→DWORD)
- NanoSpec/MainFrm.cpp: 2 instances (int→size_t)
- NanoSpec/PointConfirmationDlg.cpp: 1 instance (int→size_t)
- NanoSpec/PointSetupTeaching2PointDlg.cpp: 1 instance (int→size_t)
- NanoSpec/PointSetupTeachingDlg.cpp: 2 instances (int→size_t)

Total: 20 C4018 warnings fixed.
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/ConfigFile/IniFile.cpp
@@ -1572,7 +1572,7 @@ void SaveNanoSpecIni(int iIniFile)
 		int nIniStressSection;
 // 2009.09.17 bagus Stress --{--
 //		for(nIniStressLine = 0;nIniStressLine < STRESS_LINES_MAX;nIniStressLine++){
-		for(nIniStressLine = 0;nIniStressLine < NanoSysData.StressConfig.dwLiftPinNumberOfLine;nIniStressLine++){
+		for(nIniStressLine = 0;(DWORD)nIniStressLine < NanoSysData.StressConfig.dwLiftPinNumberOfLine;nIniStressLine++){
 // 2009.09.17 bagus Stress --}--
 			int nIniStressValidSectionNum = 0;
 			int nIniStressValidPinNum = 0;
@@ -1610,7 +1610,7 @@ void SaveNanoSpecIni(int iIniFile)
 				WritePrivateProfileString(section,szBuf,buff,szIniFilePath);
 			}
 #else
-			for (nIniStressPin=0; nIniStressPin<NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn; nIniStressPin++) {
+			for (nIniStressPin=0; (DWORD)nIniStressPin<NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn; nIniStressPin++) {
 				sprintf(szBuf,"LiftPinLocL%dPin%d_X",nIniStressLine+1,nIniStressPin+1);
 				wsprintf(buff,"%ld", NanoSysData.StressConfig.Line[nIniStressLine].PinPos[nIniStressPin].lX);
 				WritePrivateProfileString(section,szBuf,buff,szIniFilePath);
@@ -1636,7 +1636,7 @@ void SaveNanoSpecIni(int iIniFile)
 			if (NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn < 1)
 				continue;
 
-			for(nIniStressSection = 0;nIniStressSection < NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn - 1;nIniStressSection++){
+			for(nIniStressSection = 0;(DWORD)nIniStressSection < NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn - 1;nIniStressSection++){
 // 2009.09.17 bagus SE --}--
 				sprintf(szBuf,"L%dS%d_ScanStart_X",nIniStressLine+1,nIniStressSection+1);
 				wsprintf(buff,"%ld", NanoSysData.StressConfig.Line[nIniStressLine].SectPos[nIniStressSection].lScanStartPosX);
